requirements.txt
alembic==1.12.1
bcrypt==4.0.1
blinker==1.6.3
Bootstrap-Flask==2.3.2
Brotli==1.1.0
click==8.1.7
configparser==6.0.0
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Compress==1.17
Flask-Login==0.6.3
flask-mailman==0.3.0
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-Security-Too==5.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mkdocs-material-extensions==1.3
packaging==23.2
passlib==1.7.4
pip==24.3.1
python-dateutil==2.8.2
PyMySQL==1.1.0
Setuptools==68.2.2
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
Werkzeug==3.0.1
wheel==0.41.2
WTForms==3.1.1
zstandard==0.23.0
